Cleanup cache directory after split command
The subtree cache directory was previously growing out of control.
There was no safe way to remove individual cache sessions due to
the usage of `$$`. There were no safe alternatives to remove the
subtree cache in automated environments.

One could simply `rm -rf .git/subtree-cache` but this would leave
the door open for race conditions. If another subtree split command
was running at that time it would wipe out the cache for that process
leaving it in a potentially unsafe state.